first remove some obsolete bits and ditch typescript
