+++ .travis.yml
@@ -2,14 +2,8 @@ sudo: false
 language: node_js
 node_js:
   - 8
-install:
-  - npm i -g bower
-  - npm i
-  - bower i
 before_script:
   - npm run build
 branches:
   except:
     - gh-pages
-addons:
-  sauce_connect: true
