--- bower.json
@@ -1,43 +1,0 @@
-{
-  "name": "ld-navigation",
-  "version": "0.4.0-a2",
-  "authors": [
-    "tpluscode <tpluscode@users.noreply.github.com>"
-  ],
-  "keywords": [
-    "history",
-    "routing",
-    "web-components"
-  ],
-  "moduleType": "globals",
-  "homepage": "http://t-code.pl/ld-navigation/",
-  "main": "ld-navigation.html",
-  "license": "MIT",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ],
-  "devDependencies": {
-    "webcomponentsjs": "^1.0.13",
-    "polymer": "^2.1.0",
-    "web-component-tester": "^6.2.0",
-    "app-layout": "^2.0.1",
-    "iron-doc-viewer": "^3.0.4",
-    "iron-pages": "^2.0.0",
-    "paper-item": "^2.0.0",
-    "paper-listbox": "^2.0.0",
-    "paper-toast": "^2.0.0",
-    "github-fork-ribbon-css": "^0.2.2",
-    "upper88-title": "^0.1.0",
-    "prism-element": "^2.0.1",
-    "zero-md": "^0.2.0",
-    "paper-card": "polymerelements/paper-card#^2.0.0"
-  },
-  "resolutions": {
-    "polymer": "^2.1.0",
-    "webcomponentsjs": "^v1.0.2"
-  }
-}
