--- ld-navigation.html
@@ -1,1 +1,0 @@
-<script src="ld-navigation.js"></script>
