+++ package.json
@@ -1,22 +1,14 @@
 {
-  "name": "LdNavigation",
+  "name": "ld-navigation",
   "version": "0.4.0-a2",
-  "main": "ld-navigation.js",
-  "directories": {
-    "test": "tests"
-  },
+  "main": "index.js",
+  "license": "MIT",
   "scripts": {
     "build": "rollup -c",
-    "test": "wct --plugin sauce",
-    "test-local": "wct -p -l chrome --skip-selenium-install"
+    "test": "standard && karma start karma.es5.bs.config.js",
+    "test-local": "karma start karma.config.js"
   },
   "devDependencies": {
-    "rollup": "^0.50.0",
-    "rollup-plugin-node-resolve": "^3.0.0",
-    "rollup-plugin-typescript": "^0.8.1",
-    "typescript": "^2.5.2",
-    "wct-local": "^2.0.15",
-    "wct-sauce": "^1.8.6",
-    "web-component-tester": "^6.2.0"
+    "standard": "^12.0.1"
   }
 }
