--- src/index.ts
@@ -1,9 +1,0 @@
-import './ld-link';
-import './ld-navigator';
-import LdNavigator from './LdNavigator';
-import Helpers from './LdNavigation';
-
-export {
-    Helpers,
-    LdNavigator
-}
