--- wct.conf.json
@@ -1,34 +1,0 @@
-{
-  "plugins": {
-    "sauce": {
-      "disabled": true,
-      "browsers": [
-        {
-          "browserName": "microsoftedge",
-          "platform": "Windows 10",
-          "version": ""
-        },
-        {
-          "browserName": "Chrome",
-          "platform": "Windows 8.1",
-          "version": ""
-        },
-        {
-          "browserName": "Firefox",
-          "platform": "Linux",
-          "version": ""
-        },
-        {
-          "browserName": "internet explorer",
-          "platform": "Windows 8.1",
-          "version": "11"
-        },
-        {
-          "browserName": "safari",
-          "platform": "OS X 10.11",
-          "version": "9"
-        }
-      ]
-    }
-  }
-}
